[fontc] Split main.rs and lib.rs
This moves everything but fn main() to lib.rs; we can rearrange further
from there.
